implement a special handler for broken editline implementation
The editline implementation from php is broken. It does not count the
characters (end is always 0) and the line buffer contains the last line
from beginning.

To overcome these flaws we use the input observer and just allow
deleting the last character of the current line (cursor keys will not
work).