implement drawings feature
A drawing is a text with a static number of lines that should not be
broken into smaller parts and can change its content (for example a
progress bar).

Therefore when other output is made it has to get removed and then
redrawn when the output is made. This only works in terminals and when
the output is redirected it will output the last state of a drawing
before the drawing gets removed.